Adds Interval module
This adds a module for handling an `Interval` type, which seeks to
represent the distance between two notes. It is incomplete, missing
functions that can determine the interval between `Notes` as well
as the ability to calculate the note a given interval away from a
provided note.

It's also unfortunate that in this module, we're ignoring warnings
about incomplete pattern matches. This is to prevent warnings from
`intervalToHalfSteps` from complaining because we don't really
want to fail with this function since we can ALWAYS determine the
number of half steps an interval represents, but the `Interval`
type technically allows for impossible intervals despite how it
can't really happen in practice due to our only exposing the type
constructor.