Adds Scale module
This adds a module handling the `Scale` type, which is represented
as an ADT for the different classes of scale (specifically, scales
in Western classical 12-TET tuning). Scales with 2 notes through
12 notes are represented as their own constructors, with each
wrapping a `FixedList` of length n-1, where n is the number of
notes in the scale from the tonic to the octave non-inclusive.
This allows us to prohibit unrepresentable scale lengths, but the
trade-off is that it can be somewhat inconvenient to work with the
library providing this type. To help alleviate that somewhat, we've
provided convenience functions acting as constants for common
scales. This list will likely be expanded to include other common
scales (and possibly not-so-common scales), but for now it's just
the major, natural minor, and chromatic scales.