Adds Eq instances and scale inversions
This adds `Eq` instances for the `Interval` and `Scale` types based
on the interval(s) half-steps, and a function to invert a `Scale`
by reversing and inverting all of its intervals. It was checked by
ensuring that the major scale inverted to the phrygian mode, the
minor scale inverted to the mixolydian mode, and the chromatic
scale inverted to itself.