Adds enharmonic equivalence, separates Pitch and Note
This introduces `Note`, which is a broader representation of a
musical note that includes both sounding notes and `Rest`s. They
both include a `Duration`, with `Note` also including a `Pitch`
(the new name for what was previously called Note`) as well as the
note's `Octave`.

`Duration` is a representation of note duration using a `Ratio` of
`Natural` - that is, an always positive "improper fraction" which
represents the amount of a whole note it takes up. This makes the
most common note types (which have helper constants defined for
them) easy to understand, while still offering support for more
complex rhythmic ideas. Time signatures can be represented in a
similar way (and will be when they are introduced next); this will
allow support for "irrational" time signatures as well.

Finally, enharmonic equivalence is introduced using a simple
conversion to an `Int` value to make the comparison.