Adds Time module for Duration, Meter, and Tempo
This adds a new module that handles all things relating to Time.
`Duration` has been moved into it, and two more types were added:
`Tempo`, a representation of musical speed; `Meter`, or the
measurement of beats in a measure of music. Included were some
functions, constructor operators, and type aliases for convenience.