Adds documentation and convenience functions to Scale modules
This adds more Haddock documentation to `Data.Music.Scale` and
`Data.Music.Scales`, as well as more helper functions to do things
like add `Interval`s to a `Scale`, and find the union, difference,
and intersection of `Scale`s. We're also now attempting to dedupe
`Scale`s when building them, and as a result, there should be no
chance of failure since a complete list of intervals that are all
a different number of half-steps from the tonic would be the
chromatic scale. This is due to a bug in the `Ord` instance of
`Interval` which will be fixed in the next commit.