Fixes Ord instance for Interval
This changes the `Ord` instance for `Interval` to first check for
enharmonic equivalence between the two `Interval`s and return `True`
if they are. This allows us to rely on `Set` to efficiently dedupe
and find the union, difference, and intersection of a given
collection of `Interval`s. Because this means it's impossible for a
`Scale` to be larger than chromatic (11 `Intervals` excluding the
unison), we can allow `scaleFromIntervals`, `addInterval`, and
`mergeScales` to safely return without the possibility of failure.