Adds failure case for Scale construction back
This reintroduces the `Either` return for `scaleFromIntervals`,
`addInterval`, and `mergeScales`. It is possible that a user could
pass a list of tonic-equivalent `Interval`s (unison and/or octave)
to these functions which would cause it to fail. Since that's
an inescapable situation, we force the user to handle the failure
case. Realistically, `mergeScales` should never fail since it's
built from two `Scale` so it can't be impossible states, but there
isn't a way to recombine the `Scale`s without effectively taking
`scaleFromIntervals` and duplicating it unsafely while relying on
the knowledge that it can't really have a bad state. I don't know
the best solution, but since it's using `scaleFromIntervals` now,
I figure it should also return a failure case.