Adds unsafe scale construction
This introduces a function `unsafeMakeScale` which blindly attempts
to build a scale without regard for empty list arguments or lists
larger than 11 intervals. `scaleFromInterval` still fails with
`Either` but this allows us to not fail with `addInterval` or
`mergeScales` since we know these shouldn't actually have a chance
of failure, but had to account for it since they were using
`scaleFromInterval`. This function is exposed for users that know
what they're doing and want to build a `Scale` in an unsafe but
educated way without having to expose the `FixedList` constructor.
Finally, we use `unsafeMakeScale` in `Data.Music.Scales` as well.